+++ include/core/parameters.h
@@ -1864,6 +1864,12 @@ namespace Parameters
       {
         return vector_of_string;
       }
+
+    AssertThrow(false,
+                ExcMessage("Error: convert_string_to_vector only works for "
+                           "int, double and string types."));
+
+    return std::vector<T>();
   }
 
 } // namespace Parameters