+++ source/solvers/output_struct.cc
@@ -11,20 +11,25 @@ template struct OutputStructPostprocessor<2, GlobalVectorType>;
 template struct OutputStructPostprocessor<3, GlobalVectorType>;
 template struct OutputStructPostprocessor<2, GlobalBlockVectorType>;
 template struct OutputStructPostprocessor<3, GlobalBlockVectorType>;
+#ifndef LETHE_USE_LDV
 template struct OutputStructPostprocessor<
   2,
   LinearAlgebra::distributed::Vector<double>>;
 template struct OutputStructPostprocessor<
   3,
   LinearAlgebra::distributed::Vector<double>>;
+#endif
 
 template struct OutputStructSolution<2, GlobalVectorType>;
 template struct OutputStructSolution<3, GlobalVectorType>;
 template struct OutputStructSolution<2, GlobalBlockVectorType>;
 template struct OutputStructSolution<3, GlobalBlockVectorType>;
+
+#ifndef LETHE_USE_LDV
 template struct OutputStructSolution<
   2,
   LinearAlgebra::distributed::Vector<double>>;
-template struct OutputStructDoFHandler<
+template struct OutputStructSolution<
   3,
   LinearAlgebra::distributed::Vector<double>>;
+#endif